PR #4201 applied to smart_holder branch.
